Extract chat orchestration out of submissions.ts
Phase 3 Wave B, batch 4: pulls the mini chat-agent cluster —
buildChatAgentStatus, recentChatTurns, runChatAgent, and
relayedMessageLocalization — into apps/api/src/creation/chat-orchestration.ts
as createChatOrchestration(), a plain factory (no Fastify routes of its
own, same shape as the batch-3 build-status extraction).

The cluster owns everything it touches that nothing else in submissions.ts
reads: the VertexStudioChatAgent instance, its logger, the per-IP chat
rate-limit bucket, and the translator (moved wholesale, including the
2026-08-04 billed-Vertex-call incident note that justified bringing it
back). submissions.ts now only resolves chatGate/dailyChatQuota (still
shared machinery) and passes them, plus the constructor overrides from
SubmissionRoutesOptions, into the factory.

Two call sites needed rewiring: the feedback and improve routes call
runChatAgent directly; startImprovementRound and continueDraftRound
(both still central, calling dispatchBuild/resumeBuild) now import
relayedMessageLocalization from the new module instead of a local
closure. Also fixes an orphaned doc comment: startImprovementRound's
JSDoc had drifted above the chat cluster over past extractions and is
now back where it belongs.

submissions.ts drops from 5197 to 5015 lines with no behavior change.
